requirements.txt
flask==2.0.1
requests==2.26.0
gql==3.0.0
python-dotenv==0.19.0
Flask
urllib3 >= 1.15.1
intuit-oauth==1.2.4
requests_oauthlib>=1.3.1
asana
python-quickbooks
pprintpp
config
jwt
graphql-query
python_graphql_client
gql[all]
python-quickbooks

